refactor(SessionSidebar): 将叶子子组件拆分到独立文件
将 PathLabel、AnimatedDropdown、PiAgentTitle（含 useScramble），以及
session-tree 单元（SessionTreeItem/SessionItem/indicators）从单体文件中
抽取到 components/session-sidebar/*。保持行为不变，外壳再将它们导入回来。

SessionSidebar.tsx：2023 -> 1318 行。下一步：拆解主体的两个下拉区域，
使其降到 1000 行以下。
