+++ components/session-sidebar/AnimatedDropdown.tsx
@@ -1,0 +1,48 @@
+"use client";
+
+import { useState, useEffect, type CSSProperties, type ReactNode } from "react";
+
+export const DROPDOWN_ANIMATION_MS = 140;
+
+export function AnimatedDropdown({ open, children, style }: { open: boolean; children: ReactNode; style: CSSProperties }) {
+  const [mounted, setMounted] = useState(open);
+  const [visible, setVisible] = useState(open);
+
+  useEffect(() => {
+    let frame: number | undefined;
+    let timeout: ReturnType<typeof setTimeout> | undefined;
+
+    if (open) {
+      setMounted(true);
+      setVisible(false);
+      frame = window.requestAnimationFrame(() => {
+        frame = window.requestAnimationFrame(() => setVisible(true));
+      });
+    } else {
+      setVisible(false);
+      timeout = setTimeout(() => setMounted(false), DROPDOWN_ANIMATION_MS);
+    }
+
+    return () => {
+      if (frame !== undefined) window.cancelAnimationFrame(frame);
+      if (timeout) clearTimeout(timeout);
+    };
+  }, [open]);
+
+  if (!mounted) return null;
+
+  return (
+    <div
+      style={{
+        ...style,
+        opacity: visible ? 1 : 0,
+        transform: visible ? "translateY(0) scale(1)" : "translateY(-8px) scale(0.96)",
+        transformOrigin: "top center",
+        transition: `opacity ${DROPDOWN_ANIMATION_MS}ms ease, transform ${DROPDOWN_ANIMATION_MS}ms ease`,
+        pointerEvents: open ? "auto" : "none",
+      }}
+    >
+      {children}
+    </div>
+  );
+}
