+++ components/session-sidebar/PathLabel.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import type { CSSProperties } from "react";
+
+/**
+ * Path label that ellipsizes on the LEFT, keeping the (most relevant) trailing
+ * segments visible: "…orkspace/pi-web". Shows as much of the path as fits
+ * instead of a fixed number of segments. The rtl container moves the ellipsis
+ * to the left edge; the inner plaintext bidi isolation keeps the path itself
+ * rendered strictly left-to-right (no punctuation reordering).
+ */
+export function PathLabel({ text, style }: { text: string; style?: CSSProperties }) {
+  return (
+    <span
+      style={{
+        overflow: "hidden",
+        textOverflow: "ellipsis",
+        whiteSpace: "nowrap",
+        display: "block",
+        minWidth: 0,
+        lineHeight: 1.35,
+        direction: "rtl",
+        textAlign: "left",
+        ...style,
+      }}
+    >
+      <span style={{ unicodeBidi: "plaintext" }}>{text}</span>
+    </span>
+  );
+}
